@@ -266,7 +266,7 @@ class SmoothModalSheetHeader extends StatelessWidget implements SizeWidget {
           color: backgroundColor ?? primaryColor,
           boxShadow: <BoxShadow>[
             BoxShadow(
-              color: Colors.black.withOpacity(0.2),
+              color: Colors.black.withValues(alpha: 0.2),
               blurRadius: 5.0,
             ),
           ],
