@@ -25,7 +25,7 @@ Future<int?> openQuestionPage(
 }) =>
     showGeneralDialog<int?>(
       context: context,
-      barrierColor: Colors.black.withOpacity(0.7),
+      barrierColor: Colors.black.withValues(alpha: 0.7),
       pageBuilder: (_, __, ___) => EMPTY_WIDGET,
       transitionBuilder: (
         BuildContext context,
